refactoring of Program.cs
- changed User specific code to reside in a seperate class.
