+++ Chirp.CLI/cheepDB.csv
@@ -6,3 +6,4 @@ ropf,"Cheeping cheeps on Chirp :)",1690981487
 vmem,"test hej, med dig",1725714223
 vmem,"hej, med, dig",1725714235
 vmem,ni hao deluxe edition,1725720775
+vmem,test test test deluxe,1725985298
